# internal/api/docs/swagger.yaml
components:
  schemas:
...
    api.apiError:
      properties:
        code:
          enum:
          - invalid_json
          - unsupported_media_type
          - not_logged_in
          - session_expired
          - admin_only
          - missing_profile
          - unknown_profile
          - internal
          - not_found
          - invalid_credentials
          - already_set_up
          - username_required
          - password_length
          - unsupported_language
          - wrong_password
          - name_required
          - unknown_avatar
          - profile_limit
          - last_profile
          - username_taken
          - delete_self
          - invalid_id
          - invalid_request
          - episode_required
          - tmdb_unreachable
          - invalid_link
          - link_dead
          - invalid_progress
          - flaresolverr_unreachable
          - device_code_expired
          - device_code_invalid
          - release_unreachable
          - too_many_attempts
          type: string
        error:
          type: string
      required:
      - code
      - error
      type: object
...
    api.deviceApproval:
      properties:
        code:
          type: string
      required:
      - code
      type: object
...
    api.flareSolverrSettings:
      properties:
        url:
          description: URL of the FlareSolverr server, "" when disabled.
          type: string
      required:
      - url
      type: object
...
    api.instanceUpdate:
      properties:
        language:
          enum:
          - en
          - fr
          type: string
      required:
      - language
      type: object
...
  securitySchemes:
    SessionCookie:
      in: header
      name: X-Profile-ID
      type: apiKey
...
info:
  description: Self-hosted multi-source streaming (Go port of vStream).
  title: kinora API
  version: "1.0"
openapi: 3.1.0
paths:
  /admin/flaresolverr:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.flareSolverrSettings'
          description: OK
      summary: FlareSolverr server used for sources behind Cloudflare
      tags:
      - admin
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.flareSolverrSettings'
              description: FlareSolverr URL
              summary: body
        description: FlareSolverr URL
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Request
      summary: Set the FlareSolverr server (empty URL disables it), checked before
        saving. Sources behind Cloudflare are queried while it is set.
      tags:
      - admin
  /admin/instance:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.instanceUpdate'
              description: Language
              summary: body
        description: Language
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Request
      summary: Change the instance language (catalog, UI and sources)
      tags:
      - admin
  /admin/update/check:
    post:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.release'
          description: OK
        "204":
          description: No release available
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Gateway
      summary: Check for a new TV app release now, skipping the cache
      tags:
      - admin
  /admin/users:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.User'
                type: array
          description: OK
      summary: List accounts
      tags:
      - admin
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.newUser'
              description: Account
              summary: body
        description: Account
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.User'
          description: Created
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Create an account
      tags:
      - admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Delete an account
      tags:
      - admin
    patch:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.userUpdate'
              description: Settings
              summary: body
        description: Settings
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Not Found
      summary: Update an account. A new stream limit applies from the next title played.
      tags:
      - admin
...
  /auth/device/approve:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.deviceApproval'
              description: Code shown on the device
              summary: body
        description: Code shown on the device
        required: true
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Unknown or expired code
      summary: Sign a device in with the code it shows
      tags:
      - auth
  /auth/device/poll:
    post:
      description: 'Not approved yet: approved is false. Approved: the session cookie
        is set, like /auth/login.'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.devicePoll'
              description: Token from /auth/device
              summary: body
        description: Token from /auth/device
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.devicePollResult'
          description: OK
        "410":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: 'Expired: start over'
      summary: Poll a device sign-in
      tags:
      - auth
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.credentials'
              description: Credentials
              summary: body
        description: Credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.User'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Unauthorized
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: 'Too many failed attempts from this IP: retry after Retry-After
            seconds'
      summary: Log in
      tags:
      - auth
...
  /catalog/discover:
    get:
      parameters:
      - description: movie or tv
        in: query
        name: type
        required: true
        schema:
          type: string
      - description: TMDB genre ID
        in: query
        name: genre
        required: true
        schema:
          type: integer
      - description: Page, from 1
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/tmdb.Item'
                type: array
          description: OK
      summary: Titles of a genre, most popular first
      tags:
      - catalog
  /catalog/foryou:
    get:
      parameters:
      - description: movie or tv, empty for both
        in: query
        name: type
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.row'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: 'Rows personalized for the profile: "Because you watched X", from TMDB
        recommendations of recent titles'
      tags:
      - catalog
  /catalog/genres:
    get:
      parameters:
      - description: movie or tv
        in: query
        name: type
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/tmdb.Genre'
                type: array
          description: OK
      summary: Genres of movies or shows
      tags:
      - catalog
  /catalog/home:
    get:
      parameters:
      - description: movie or tv, empty for both
        in: query
        name: type
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.row'
                type: array
          description: OK
      summary: Home rows (all, movies or shows)
      tags:
      - catalog
  /catalog/search:
    get:
      parameters:
      - description: Query
        in: query
        name: q
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/tmdb.Item'
                type: array
          description: OK
      summary: Search movies and shows
      tags:
      - catalog
...
  /library/family:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.ListItem'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: '"Family list", shared by every profile of the account, without the
        titles the profile gave a thumbs down'
      tags:
      - library
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/store.ListItem'
              description: Title
              summary: body
        description: Title
        required: true
      responses:
        "204":
          description: No Content
      security:
      - SessionCookie: []
      summary: Add a title to the "Family list"
      tags:
      - library
  /library/family/{type}/{id}:
    delete:
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - SessionCookie: []
      summary: Remove a title from the "Family list"
      tags:
      - library
  /library/finished:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.ListItem'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Movies watched to the end and shows watched up to their latest aired
        episode, most recent first
      tags:
      - library
  /library/history:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.HistoryEntry'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Last movies and episodes the profile played, most recent first
      tags:
      - library
  /library/list:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.ListItem'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: '"My list" of the profile'
      tags:
      - library
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/store.ListItem'
              description: Title
              summary: body
        description: Title
        required: true
      responses:
        "204":
          description: No Content
      security:
      - ProfileHeader: []
      summary: Add a title to "My list"
      tags:
      - library
  /library/list/{type}/{id}:
    delete:
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - ProfileHeader: []
      summary: Remove a title from "My list"
      tags:
      - library
  /library/progress:
    get:
      description: |-
        Titles in progress and the next episode of the shows whose last watched episode is finished, most
        recent first. A next episode the profile caught up with (badge) goes first if it aired in the last
        month, last otherwise.
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.Progress'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: '"Continue watching" row of the profile'
      tags:
      - library
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/store.Progress'
              description: Position
              summary: body
        description: Position
        required: true
      responses:
        "204":
          description: No Content
      security:
      - ProfileHeader: []
      summary: Save the playback position
      tags:
      - library
  /library/progress/{type}/{id}:
    get:
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.Progress'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Watch progress of a title (every episode for a show), most recent first
      tags:
      - library
  /library/ratings:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.RatedItem'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Thumbs up and down of the profile, most recent first
      tags:
      - library
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/store.RatedItem'
              description: Title and rating
              summary: body
        description: Title and rating
        required: true
      responses:
        "204":
          description: No Content
      security:
      - ProfileHeader: []
      summary: Thumbs up (1) or down (-1) on a title, 0 removes it
      tags:
      - library
  /library/stats:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.YearStats'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Time spent watching per year, most recent first
      tags:
      - library
  /library/watched:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.Watched'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Every title the profile played, finished or not, most recent first
      tags:
      - library
  /library/watched/{type}/{id}:
    post:
      description: A movie, or every episode of a show aired so far. Titles already
        finished keep their date.
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      security:
      - ProfileHeader: []
      summary: Mark a title as entirely watched
      tags:
      - library
...
  /me/password:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.passwordChange'
              description: Passwords
              summary: body
        description: Passwords
        required: true
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Unauthorized
      summary: Change the account password (logs out every session)
      tags:
      - auth
  /play:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.playRequest'
              description: Link token
              summary: body
        description: Link token
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.playResponse'
          description: OK
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Gateway
      summary: Resolve a link into a proxied stream URL
      tags:
      - playback
  /profiles:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.Profile'
                type: array
          description: OK
      summary: Profiles of the account
      tags:
      - profiles
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.profileBody'
              description: Profile
              summary: body
        description: Profile
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.Profile'
          description: Created
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Create a profile (5 max per account)
      tags:
      - profiles
  /profiles/{id}:
    delete:
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      summary: Delete a profile and its history
      tags:
      - profiles
    patch:
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.profileBody'
              description: Profile
              summary: body
        description: Profile
        required: true
      responses:
        "204":
          description: No Content
      summary: Rename or change the avatar of a profile
      tags:
      - profiles
  /setup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.setupRequest'
              description: Admin credentials and language
              summary: body
        description: Admin credentials and language
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.User'
          description: Created
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Create the first admin account and pick the instance language (only
        while no account exists), then log in
      tags:
      - auth
  /titles/{type}/{id}:
    get:
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tmdb.Details'
          description: OK
      summary: Details of a movie or show
      tags:
      - catalog
  /titles/{type}/{id}/links:
    get:
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: Season (shows)
        in: query
        name: season
        schema:
          type: integer
      - description: Episode (shows)
        in: query
        name: episode
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.link'
                type: array
          description: OK
      summary: Find links for a movie or an episode across the sources of the instance
        language
      tags:
      - playback
  /titles/{type}/{id}/segments:
    get:
      parameters:
      - description: movie or tv
        in: path
        name: type
        required: true
        schema:
          type: string
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: Season (shows)
        in: query
        name: season
        schema:
          type: integer
      - description: Episode (shows)
        in: query
        name: episode
        schema:
          type: integer
      - description: Length of the file being played, in seconds
        in: query
        name: duration
        required: true
        schema:
          type: number
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/aniskip.Segment'
                type: array
          description: OK
      security:
      - ProfileHeader: []
      summary: Opening and ending of a movie or episode, for the skip buttons (empty
        when unknown or turned off for the profile)
      tags:
      - catalog
  /titles/tv/{id}/seasons/{season}:
    get:
      parameters:
      - description: TMDB ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      - description: Season number
        in: path
        name: season
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/tmdb.Episode'
                type: array
          description: OK
      summary: Episodes of a season
      tags:
      - catalog
...
servers:
- url: /api/v1